DS-273: Created link directive
